requirements.txt
asgiref==3.11.1
Django==3.2.5
django-cors-headers==4.5.0
djangorestframework==3.15.1
gunicorn==25.1.0
packaging==26.0
psycopg2-binary==2.9.11
python-dotenv==1.2.2
pytz==2026.1.post1
sqlparse==0.5.5
tzdata==2025.3
